# apps/cli/openapi.yaml
...
openapi: "3.1.0"
info:
  title: agent-media Edge Functions API
  version: "0.1.0"
  description: >-
    RESTful API for AI-powered video and image generation. All endpoints are
    hosted as Supabase Edge Functions. Authenticated endpoints accept a Bearer
    JWT (Supabase auth token) or an API key prefixed with `am_`.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: agent-media
    url: https://agent-media.ai
servers:
  - url: https://ppwvarkmpffljlqxkjux.supabase.co
    description: Production Supabase project

# ──────────────────────────────────────────────────────────────────────────────
# Security Schemes
# ──────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT access token obtained via OTP or OAuth login.

    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: >-
        API key issued by the apikey-manage endpoint. Pass as
        `Bearer am_<key>`. The server detects the `am_` prefix and validates
        against the hashed key store.

  # ────────────────────────────────────────────────────────────────────────────
  # Reusable Schemas
  # ────────────────────────────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code.
          examples: ["unauthorized", "invalid_request", "server_error"]
        error_description:
          type: string
          description: Human-readable explanation.
    RateLimitError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          properties:
            retry_after:
              type: integer
              description: Seconds until the rate limit window resets.

    # ── Device Token ──────────────────────────────────────────────────────
...
    GenerateRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
          description: Model slug, e.g. "kling3", "flux2-pro".
        prompt:
          type: string
          maxLength: 4000
        duration:
          type: integer
          minimum: 1
          maximum: 30
          nullable: true
          description: Video duration in seconds.
        resolution:
          type: string
          enum: ["360p", "480p", "720p", "1080p", "2k", "4k"]
          nullable: true
        aspect_ratio:
          type: string
          enum: ["1:1", "4:3", "3:4", "16:9", "9:16", "21:9"]
          nullable: true
        seed:
          type: integer
          minimum: 0
          nullable: true
        input_media_url:
          type: string
          format: uri
          nullable: true
          description: URL or storage:// path for image-to-video workflows.
        idempotency_key:
          type: string
          nullable: true
        dry_run:
          type: boolean
          default: false
          description: When true, returns cost estimate without submitting.
    GenerateResponse:
      type: object
      required: [job_id, status, estimated_duration, credits_deducted]
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [submitted]
        estimated_duration:
          type: integer
          description: Estimated processing time in seconds.
        credits_deducted:
          type: integer
    GenerateDryRunResponse:
      type: object
      required: [credits, costUsd, availableCredits, canAfford]
      properties:
        credits:
          type: integer
        costUsd:
          type: number
        availableCredits:
          type: integer
        canAfford:
          type: boolean
        modelDisplayName:
          type: string
        breakdown:
          type: object
          properties:
            planCredits:
              type: integer
            purchasedCredits:
              type: integer
    InsufficientCreditsError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          properties:
            credits_required:
              type: integer
            credits_available:
              type: integer
            monthly_remaining:
              type: integer
            purchased_balance:
              type: integer

    # ── Job Status ─────────────────────────────────────────────────────────
...
    PricingEntry:
      type: object
      required: [modelSlug, operation, creditCost, providerCostUsd]
      properties:
        modelSlug:
          type: string
        operation:
          type: string
          enum: [text_to_video, image_to_video, text_to_image]
        durationSeconds:
          type: integer
          nullable: true
        resolution:
          type: string
          nullable: true
        creditCost:
          type: integer
        providerCostUsd:
          type: number
    PricingResponse:
      type: object
      required: [pricing]
      properties:
        pricing:
          type: array
          items:
            $ref: "#/components/schemas/PricingEntry"

    # ── Upload URL ─────────────────────────────────────────────────────────
...
paths:
...
  /functions/v1/generate:
    post:
      operationId: generate
      summary: Submit a media generation job
      description: >-
        11-step pipeline: authenticate, validate, lookup model, check circuit
        breaker, check tier and concurrent limits, calculate cost, handle
        idempotency, deduct credits atomically, resolve media URLs, submit to
        provider, create job record.
      tags: [Generation]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "201":
          description: Job submitted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "200":
          description: Dry-run cost estimate (when dry_run=true).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateDryRunResponse"
        "400":
          description: Invalid input or content moderation failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: Insufficient credits.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientCreditsError"
        "403":
          description: Model not available on plan or feature disabled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Duplicate idempotency key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "429":
          description: Rate limit or concurrent job limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
        "500":
          description: Provider error (credits refunded).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      credits_refunded:
                        type: integer
        "503":
          description: Provider temporarily unavailable (circuit breaker open).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      retry_after:
                        type: integer
                      provider_slug:
                        type: string

  # ── Job Status ───────────────────────────────────────────────────────────
...
  /functions/v1/pricing:
    get:
      operationId: pricing
      summary: Get model pricing
      description: >-
        Returns credit costs and USD prices for all active models. Public
        endpoint; no authentication required. Supports optional model filter.
      tags: [Billing]
      parameters:
        - name: model
          in: query
          required: false
          schema:
            type: string
          description: Filter pricing to a specific model slug.
      responses:
        "200":
          description: Pricing data returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingResponse"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitError"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Upload URL ───────────────────────────────────────────────────────────
...